Make CandidateNotesTest and CandidateApprovalTest extend CandidateTest
