+++ nvi-commons/src/test/java/no/sikt/nva/nvi/common/service/CandidateTest.java
@@ -24,14 +24,12 @@ import static nva.commons.core.ioutils.IoUtils.stringFromResources;
 import static nva.commons.core.paths.UriWrapper.HTTPS;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.not;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.hamcrest.Matchers.nullValue;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -41,7 +39,6 @@ import java.math.RoundingMode;
 import java.net.URI;
 import java.nio.file.Path;
 import java.time.ZonedDateTime;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -62,8 +59,6 @@ import no.sikt.nva.nvi.common.db.PeriodStatus;
 import no.sikt.nva.nvi.common.db.PeriodStatus.Status;
 import no.sikt.nva.nvi.common.db.ReportStatus;
 import no.sikt.nva.nvi.common.db.model.ChannelType;
-import no.sikt.nva.nvi.common.model.UpdateAssigneeRequest;
-import no.sikt.nva.nvi.common.model.UpdateStatusRequest;
 import no.sikt.nva.nvi.common.service.dto.ApprovalDto;
 import no.sikt.nva.nvi.common.service.dto.ApprovalStatusDto;
 import no.sikt.nva.nvi.common.service.dto.PeriodStatusDto;
@@ -75,8 +70,6 @@ import no.sikt.nva.nvi.common.service.model.GlobalApprovalStatus;
 import no.sikt.nva.nvi.common.service.model.InstanceType;
 import no.sikt.nva.nvi.common.service.model.InstitutionPoints;
 import no.sikt.nva.nvi.common.service.model.InstitutionPoints.CreatorAffiliationPoints;
-import no.sikt.nva.nvi.common.service.model.PublicationChannel;
-import no.sikt.nva.nvi.common.service.model.PublicationDetails.Creator;
 import no.sikt.nva.nvi.common.service.model.PublicationDetails.PublicationDate;
 import no.sikt.nva.nvi.common.service.requests.UpsertCandidateRequest;
 import no.sikt.nva.nvi.test.LocalDynamoTest;
@@ -84,7 +77,6 @@ import nva.commons.core.Environment;
 import nva.commons.core.paths.UriWrapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Named;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
@@ -96,22 +88,30 @@ import org.junit.jupiter.params.provider.ValueSource;
 @SuppressWarnings({"PMD.GodClass", "PMD.CouplingBetweenObjects"})
 class CandidateTest extends LocalDynamoTest {
 
+    protected static final int EXPECTED_SCALE = 4;
+    protected static final RoundingMode EXPECTED_ROUNDING_MODE = RoundingMode.HALF_UP;
     private static final Environment ENVIRONMENT = new Environment();
     private static final String BASE_PATH = ENVIRONMENT.readEnv("CUSTOM_DOMAIN_BASE_PATH");
     private static final String API_DOMAIN = ENVIRONMENT.readEnv("API_HOST");
     public static final URI CONTEXT_URI = UriWrapper.fromHost(API_DOMAIN).addChild(BASE_PATH, "context").getUri();
-    private static final int EXPECTED_SCALE = 4;
-    private static final RoundingMode EXPECTED_ROUNDING_MODE = RoundingMode.HALF_UP;
-    private static final URI HARDCODED_INSTITUTION_ID = URI.create("https://example.org/topLevelInstitutionId");
-    private static final URI HARDCODED_SUBUNIT_ID = URI.create("https://example.org/subUnitInstitutionId");
-    private static final URI HARDCODED_CHANNEL_ID = URI.create(
-        "https://example.org/publication-channels-v2/journal/123/2018");
-    private static final String HARDCODED_LEVEL = "LevelOne";
-    private static final URI HARDCODED_CREATOR_ID = URI.create("https://example.org/someCreator");
-    private static final InstanceType HARDCODED_INSTANCE_TYPE = InstanceType.ACADEMIC_ARTICLE;
-    private static final BigDecimal HARDCODED_POINTS = BigDecimal.ONE.setScale(EXPECTED_SCALE, EXPECTED_ROUNDING_MODE);
-    private CandidateRepository candidateRepository;
-    private PeriodRepository periodRepository;
+    protected CandidateRepository candidateRepository;
+    protected PeriodRepository periodRepository;
+
+    protected static UpsertCandidateRequest createUpsertRequestWithDecimalScale(int scale, URI institutionId) {
+        var creatorId = randomUri();
+        var creators = Map.of(creatorId, List.of(institutionId));
+        var points = List.of(createInstitutionPoints(institutionId, randomBigDecimal(scale), creatorId));
+
+        return createUpsertCandidateRequest(randomUri(), randomUri(), true,
+                                            new PublicationDate(String.valueOf(CURRENT_YEAR), null, null),
+                                            creators,
+                                            randomInstanceType(),
+                                            randomElement(ChannelType.values()).getValue(), randomUri(),
+                                            randomLevelExcluding(DbLevel.NON_CANDIDATE).getValue(), points,
+                                            randomInteger(), randomBoolean(),
+                                            randomBigDecimal(scale), randomBigDecimal(scale),
+                                            randomBigDecimal(scale));
+    }
 
     @BeforeEach
     void setup() {
@@ -401,115 +401,6 @@ class CandidateTest extends LocalDynamoTest {
     }
 
     @Test
-    void shouldNotOverrideAssigneeWhenAssigneeAlreadyIsSet() {
-        var institutionId = randomUri();
-        var assignee = randomString();
-        var request = createUpsertCandidateRequest(institutionId);
-        Candidate.upsert(request, candidateRepository, periodRepository);
-        var candidate = Candidate.fetchByPublicationId(request::publicationId, candidateRepository, periodRepository)
-                            .updateApproval(new UpdateAssigneeRequest(institutionId, assignee))
-                            .updateApproval(
-                                createUpdateStatusRequest(ApprovalStatus.APPROVED, institutionId, randomString()))
-                            .updateApproval(
-                                createUpdateStatusRequest(ApprovalStatus.REJECTED, institutionId, randomString()))
-                            .toDto();
-        assertThat(candidate.approvals().getFirst().assignee(), is(equalTo(assignee)));
-        assertThat(candidate.approvals().getFirst().finalizedBy(), is(not(equalTo(assignee))));
-    }
-
-    @Test
-    void shouldNotResetApprovalsWhenUpdatingCandidateFieldsNotEffectingApprovals() {
-        var institutionId = randomUri();
-        var upsertCandidateRequest = createUpsertCandidateRequest(institutionId);
-        var candidate = upsert(upsertCandidateRequest);
-        candidate.updateApproval(
-            new UpdateStatusRequest(institutionId, ApprovalStatus.APPROVED, randomString(), randomString()));
-        var approval = candidate.getApprovals().get(institutionId);
-        var newUpsertRequest = createNewUpsertRequestNotAffectingApprovals(upsertCandidateRequest);
-        var updatedCandidate = upsert(newUpsertRequest);
-        var updatedApproval = updatedCandidate.getApprovals().get(institutionId);
-
-        assertThat(updatedApproval, is(equalTo(approval)));
-    }
-
-    @Test
-    void shouldNotResetApprovalsWhenCreatorAffiliationChangesWithinSameInstitution() {
-        var upsertCandidateRequest = getUpsertCandidateRequestWithHardcodedValues();
-        var candidate = upsert(upsertCandidateRequest);
-        candidate.updateApproval(
-            new UpdateStatusRequest(HARDCODED_INSTITUTION_ID, ApprovalStatus.APPROVED, randomString(), randomString()));
-        var newSubUnitInSameOrganization = randomUri();
-        var newUpsertRequest = createUpsertCandidateRequest(candidate.getPublicationId(),
-                                                            candidate.getPublicationDetails().publicationBucketUri(),
-                                                            true,
-                                                            candidate.getPublicationDetails().publicationDate(),
-                                                            Map.of(HARDCODED_CREATOR_ID, List.of(
-                                                                newSubUnitInSameOrganization)),
-                                                            InstanceType.parse(candidate.getInstanceType()),
-                                                            candidate.getPublicationChannelType().getValue(),
-                                                            candidate.getPublicationChannelId(),
-                                                            candidate.getScientificLevel(),
-                                                            List.of(new InstitutionPoints(HARDCODED_INSTITUTION_ID,
-                                                                                          HARDCODED_POINTS, List.of(
-                                                                new CreatorAffiliationPoints(HARDCODED_CREATOR_ID,
-                                                                                             newSubUnitInSameOrganization,
-                                                                                             HARDCODED_POINTS)))),
-                                                            randomInteger(), false,
-                                                            randomBigDecimal(), null, randomBigDecimal());
-        var updatedCandidate = upsert(newUpsertRequest);
-        var updatedApproval = updatedCandidate.getApprovals().get(HARDCODED_INSTITUTION_ID);
-
-        assertThat(updatedApproval.getStatus(), is(equalTo(ApprovalStatus.APPROVED)));
-    }
-
-    @Test
-    void shouldNotResetApprovalsWhenUpsertRequestContainsSameDecimalsWithAnotherScale() {
-        var institutionId = randomUri();
-        var upsertCandidateRequest = createUpsertRequestWithDecimalScale(0, institutionId);
-        var candidate = upsert(upsertCandidateRequest);
-        candidate.updateApproval(
-            new UpdateStatusRequest(institutionId, ApprovalStatus.APPROVED, randomString(), randomString()));
-        var approval = candidate.getApprovals().get(institutionId);
-        var newUpsertRequest = createNewUpsertRequestNotAffectingApprovals(upsertCandidateRequest);
-        var updatedCandidate = upsert(newUpsertRequest);
-        var updatedApproval = updatedCandidate.getApprovals().get(institutionId);
-
-        assertThat(updatedApproval, is(equalTo(approval)));
-    }
-
-    @Test
-    void shouldResetApprovalsWhenNonCandidateBecomesCandidate() {
-        var institutionId = randomUri();
-        var upsertCandidateRequest = createUpsertCandidateRequest(institutionId);
-        var candidate = upsert(upsertCandidateRequest);
-        var nonCandidate = Candidate.updateNonCandidate(
-            createUpsertNonCandidateRequest(candidate.getPublicationId()),
-            candidateRepository).orElseThrow();
-        assertFalse(nonCandidate.isApplicable());
-        var updatedCandidate = upsert(createNewUpsertRequestNotAffectingApprovals(upsertCandidateRequest));
-
-        assertTrue(updatedCandidate.isApplicable());
-        assertThat(updatedCandidate.getApprovals().size(), is(greaterThan(0)));
-    }
-
-    @ParameterizedTest
-    @MethodSource("candidateResetCauseProvider")
-    @DisplayName("Should reset approvals when updating fields effecting approvals")
-    void shouldResetApprovalsWhenUpdatingFieldsEffectingApprovals(CandidateResetCauseArgument arguments) {
-        var upsertCandidateRequest = getUpsertCandidateRequestWithHardcodedValues();
-
-        var candidate = upsert(upsertCandidateRequest);
-        candidate.updateApproval(
-            new UpdateStatusRequest(HARDCODED_INSTITUTION_ID, ApprovalStatus.APPROVED, randomString(), randomString()));
-
-        var newUpsertRequest = getUpsertCandidateRequest(arguments, candidate.getPublicationId());
-
-        var updatedCandidate = upsert(newUpsertRequest);
-        var updatedApproval = updatedCandidate.getApprovals().get(HARDCODED_INSTITUTION_ID);
-        assertThat(updatedApproval.getStatus(), is(equalTo(ApprovalStatus.PENDING)));
-    }
-
-    @Test
     void shouldReturnNviCandidateContextAsString() {
         var expectedContext = stringFromResources(Path.of("nviCandidateContext.json"));
         assertThat(Candidate.getJsonLdContext(), is(equalTo(expectedContext)));
@@ -584,39 +475,6 @@ class CandidateTest extends LocalDynamoTest {
         assertNotEquals(dao.version(), updatedDao.version());
     }
 
-    private static Stream<Arguments> candidateResetCauseProvider() {
-        return Stream.of(
-            Arguments.of(Named.of("channel changed",
-                                  CandidateResetCauseArgument.defaultBuilder().withChannelId(randomUri()).build())),
-            Arguments.of(Named.of("level changed",
-                                  CandidateResetCauseArgument.defaultBuilder()
-                                      .withLevel("LevelTwo")
-                                      .build())),
-            Arguments.of(Named.of("type changed",
-                                  CandidateResetCauseArgument.defaultBuilder()
-                                      .withType(InstanceType.ACADEMIC_MONOGRAPH)
-                                      .build())),
-            Arguments.of(Named.of("institution points changed",
-                                  CandidateResetCauseArgument.defaultBuilder()
-                                      .withPointsForInstitution(BigDecimal.TEN)
-                                      .build())),
-            Arguments.of(Named.of("creator changed",
-                                  CandidateResetCauseArgument.defaultBuilder()
-                                      .withCreators(
-                                          List.of(new Creator(randomUri(), List.of(HARDCODED_INSTITUTION_ID))))
-                                      .build())),
-            Arguments.of(Named.of("creator removed",
-                                  CandidateResetCauseArgument.defaultBuilder()
-                                      .withCreators(Collections.emptyList())
-                                      .build())),
-            Arguments.of(Named.of("creator added",
-                                  CandidateResetCauseArgument.defaultBuilder()
-                                      .withCreators(List.of(CandidateResetCauseArgument.Builder.DEFAULT_CREATOR,
-                                                            new Creator(randomUri(),
-                                                                        List.of(HARDCODED_INSTITUTION_ID))))
-                                      .build())));
-    }
-
     @Deprecated
     private static Stream<Arguments> levelValues() {
         return Stream.of(Arguments.of(DbLevel.LEVEL_ONE, "LevelOne"), Arguments.of(DbLevel.LEVEL_TWO, "LevelTwo"));
@@ -627,22 +485,6 @@ class CandidateTest extends LocalDynamoTest {
         return new InstitutionPoints(institutionId, institutionPoints,
                                      List.of(new CreatorAffiliationPoints(creatorId, institutionId,
                                                                           institutionPoints)));
-    }
-
-    private static UpsertCandidateRequest createUpsertRequestWithDecimalScale(int scale, URI institutionId) {
-        var creatorId = randomUri();
-        var creators = Map.of(creatorId, List.of(institutionId));
-        var points = List.of(createInstitutionPoints(institutionId, randomBigDecimal(scale), creatorId));
-
-        return createUpsertCandidateRequest(randomUri(), randomUri(), true,
-                                            new PublicationDate(String.valueOf(CURRENT_YEAR), null, null),
-                                            creators,
-                                            randomInstanceType(),
-                                            randomElement(ChannelType.values()).getValue(), randomUri(),
-                                            randomLevelExcluding(DbLevel.NON_CANDIDATE).getValue(), points,
-                                            randomInteger(), randomBoolean(),
-                                            randomBigDecimal(scale), randomBigDecimal(scale),
-                                            randomBigDecimal(scale));
     }
 
     private static void assertCorrectCreatorData(Map<URI, List<URI>> creators, Candidate candidate) {
@@ -695,41 +537,6 @@ class CandidateTest extends LocalDynamoTest {
         var updateRequest = createUpsertNonCandidateRequest(tempCandidate.getPublicationId());
         var candidateBO = Candidate.updateNonCandidate(updateRequest, candidateRepository).orElseThrow();
         return Candidate.fetch(candidateBO::getIdentifier, candidateRepository, periodRepository);
-    }
-
-    private UpsertCandidateRequest getUpsertCandidateRequest(CandidateResetCauseArgument arguments, URI publicationId) {
-        return createUpsertCandidateRequest(publicationId,
-                                            randomUri(), true,
-                                            new PublicationDate(String.valueOf(CURRENT_YEAR), null, null),
-                                            arguments.creators()
-                                                .stream()
-                                                .collect(Collectors.toMap(Creator::id, Creator::affiliations)),
-                                            arguments.type(),
-                                            arguments.channel().channelType().getValue(),
-                                            arguments.channel().id(),
-                                            arguments.channel().level(),
-                                            arguments.institutionPoints(),
-                                            randomInteger(), false,
-                                            randomBigDecimal(), null, randomBigDecimal());
-    }
-
-    private UpsertCandidateRequest getUpsertCandidateRequestWithHardcodedValues() {
-        return createUpsertCandidateRequest(URI.create("publicationId"), randomUri(), true,
-                                            new PublicationDate(String.valueOf(CURRENT_YEAR), null, null),
-                                            Map.of(HARDCODED_CREATOR_ID, List.of(HARDCODED_SUBUNIT_ID)),
-                                            HARDCODED_INSTANCE_TYPE,
-                                            ChannelType.JOURNAL.getValue(),
-                                            HARDCODED_CHANNEL_ID,
-                                            HARDCODED_LEVEL,
-                                            List.of(
-                                                new InstitutionPoints(HARDCODED_INSTITUTION_ID, HARDCODED_POINTS,
-                                                                      List.of(
-                                                                          new CreatorAffiliationPoints(
-                                                                              HARDCODED_CREATOR_ID,
-                                                                              HARDCODED_SUBUNIT_ID,
-                                                                              HARDCODED_POINTS)))),
-                                            randomInteger(), false,
-                                            randomBigDecimal(), null, randomBigDecimal());
     }
 
     private UpsertCandidateRequest getUpdateRequestForExistingCandidate() {
@@ -789,144 +596,5 @@ class CandidateTest extends LocalDynamoTest {
 
     private List<DbCreator> mapToDbCreators(Map<URI, List<URI>> creators) {
         return creators.entrySet().stream().map(entry -> new DbCreator(entry.getKey(), entry.getValue())).toList();
-    }
-
-    private UpsertCandidateRequest createNewUpsertRequestNotAffectingApprovals(UpsertCandidateRequest request) {
-        return new UpsertCandidateRequest() {
-            @Override
-            public URI publicationBucketUri() {
-                return request.publicationBucketUri();
-            }
-
-            @Override
-            public URI publicationId() {
-                return request.publicationId();
-            }
-
-            @Override
-            public boolean isApplicable() {
-                return true;
-            }
-
-            @Override
-            public boolean isInternationalCollaboration() {
-                return false;
-            }
-
-            @Override
-            public Map<URI, List<URI>> creators() {
-                return request.creators();
-            }
-
-            @Override
-            public String channelType() {
-                return null;
-            }
-
-            @Override
-            public URI publicationChannelId() {
-                return request.publicationChannelId();
-            }
-
-            @Override
-            public String level() {
-                return request.level();
-            }
-
-            @Override
-            public InstanceType instanceType() {
-                return request.instanceType();
-            }
-
-            @Override
-            public PublicationDate publicationDate() {
-                return request.publicationDate();
-            }
-
-            @Override
-            public int creatorShareCount() {
-                return 0;
-            }
-
-            @Override
-            public BigDecimal collaborationFactor() {
-                return null;
-            }
-
-            @Override
-            public BigDecimal basePoints() {
-                return null;
-            }
-
-            @Override
-            public List<InstitutionPoints> institutionPoints() {
-                return request.institutionPoints();
-            }
-
-            @Override
-            public BigDecimal totalPoints() {
-                return request.totalPoints();
-            }
-        };
-    }
-
-    private record CandidateResetCauseArgument(PublicationChannel channel, InstanceType type,
-                                               List<InstitutionPoints> institutionPoints, List<Creator> creators) {
-
-        private static CandidateResetCauseArgument.Builder defaultBuilder() {
-            return new Builder();
-        }
-
-        private static final class Builder {
-
-            private static final Creator DEFAULT_CREATOR = new Creator(HARDCODED_CREATOR_ID,
-                                                                       List.of(HARDCODED_SUBUNIT_ID));
-            private PublicationChannel channel = new PublicationChannel(ChannelType.JOURNAL,
-                                                                        HARDCODED_CHANNEL_ID,
-                                                                        HARDCODED_LEVEL);
-            private InstanceType type = HARDCODED_INSTANCE_TYPE;
-            private List<InstitutionPoints> institutionPoints = List.of(
-                new InstitutionPoints(HARDCODED_INSTITUTION_ID, HARDCODED_POINTS,
-                                      List.of(new CreatorAffiliationPoints(HARDCODED_CREATOR_ID,
-                                                                           HARDCODED_SUBUNIT_ID,
-                                                                           HARDCODED_POINTS))));
-            private List<Creator> creators = List.of(DEFAULT_CREATOR);
-
-            private Builder() {
-            }
-
-            private Builder withType(InstanceType type) {
-                this.type = type;
-                return this;
-            }
-
-            private Builder withCreators(List<Creator> creators) {
-                this.creators = creators;
-                return this;
-            }
-
-            private Builder withChannelId(URI publicationChannelId) {
-                this.channel = new PublicationChannel(ChannelType.JOURNAL, publicationChannelId, HARDCODED_LEVEL);
-                return this;
-            }
-
-            private Builder withLevel(String level) {
-                this.channel = new PublicationChannel(ChannelType.JOURNAL, HARDCODED_CHANNEL_ID, level);
-                return this;
-            }
-
-            private Builder withPointsForInstitution(BigDecimal institutionPoints) {
-                this.institutionPoints = List.of(new InstitutionPoints(HARDCODED_INSTITUTION_ID, institutionPoints,
-                                                                       List.of(new CreatorAffiliationPoints(
-                                                                           HARDCODED_CREATOR_ID,
-                                                                           HARDCODED_SUBUNIT_ID,
-                                                                           HARDCODED_POINTS))));
-                return this;
-            }
-
-            private CandidateResetCauseArgument build() {
-                return new CandidateResetCauseArgument(channel, type, institutionPoints, creators);
-            }
-        }
     }
 }
